refactor(calendar): split CustomCalendar into modular and reusable components
